Split out import list into common module
